requirements-dev.txt
Faker==1.0.5
flexmock==0.10.4
ipdb==0.11
model_mommy==1.6.0
names==0.3.0
pygments==2.3.1
pytest==5.0.1
pytest-django==3.5.1
pytest-faker==2.0.0
pytest-html==1.22.0
pytest-mock==1.10.4
sqlparse==0.3.0
timeout-decorator==0.4.1
